Merge aji ke master

+++ src/model/TopUp.java
@@ -10,32 +10,32 @@ package model;
  * @author Hp
  */
 public class TopUp {
-    private int nominalTopUp;
-    private int saldo;
-    
+    private double nominalTopUp;
+    private double saldo;
+  
     public TopUp(){
         
     }
 
-    public TopUp(int nominalTopUp, int saldo) {
+    public TopUp(double nominalTopUp, double saldo) {
         this.nominalTopUp = nominalTopUp;
         this.saldo = saldo;
     }
 
-    public int getNominalTopUp() {
+    public double getNominalTopUp() {
         return nominalTopUp;
     }
 
-    public void setNominalTopUp(int nominalTopUp) {
+    public void setNominalTopUp(double nominalTopUp) {
         this.nominalTopUp = nominalTopUp;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
-  
+
 }
